requirements.txt
flask
flask-socketio
gunicorn
eventlet
python-binance
sqlalchemy
pymysql
python-dotenv